[0068] 将 liii os 相关函数从 goldfish.hpp 迁移到 src/liii_os.cpp
