+++ devel/0068.md
@@ -1,0 +1,99 @@
+# [0068] 将 (liii os) 相关函数从 goldfish.hpp 迁移到 src/liii_os.cpp
+
+## 1. 相关文档
+- [0064.md](0064.md) - njson 迁移（参照模板）
+- [0066.md](0066.md) - http 迁移（参照模板）
+- [dddd.md](dddd.md) - 任务文档模板
+
+## 2. 任务相关的代码文件
+- `src/goldfish.hpp` — 删除 os 相关实现代码，保留前向声明
+- `src/liii_os.cpp`（新建）— os 相关 C++ 实现
+- `xmake.lua` — 添加新编译单元
+- `goldfish/liii/os.scm` — Scheme 层（不改动）
+
+## 3. 如何测试
+
+### 3.1 确定性测试（单元测试）
+```bash
+xmake b goldfish
+bin/gf tests/liii/os-test.scm
+bin/gf tests/liii/os/os-type-test.scm
+bin/gf tests/liii/os/os-call-test.scm
+bin/gf tests/liii/os/os-temp-dir-test.scm
+bin/gf tests/liii/os/access-test.scm
+bin/gf tests/liii/os/getenv-test.scm
+bin/gf tests/liii/os/putenv-test.scm
+bin/gf tests/liii/os/unsetenv-test.scm
+bin/gf tests/liii/os/getcwd-test.scm
+bin/gf tests/liii/os/chdir-test.scm
+bin/gf tests/liii/os/mkdir-test.scm
+bin/gf tests/liii/os/rmdir-test.scm
+bin/gf tests/liii/os/remove-test.scm
+bin/gf tests/liii/os/rename-test.scm
+bin/gf tests/liii/os/listdir-test.scm
+```
+
+### 3.2 回归测试
+```bash
+bin/gf test --changed-since=main
+```
+
+## 4. 如何提交
+
+提交前执行以下最少步骤：
+
+```bash
+bin/gf test --changed-since=main
+```
+
+## 5. 迁移方案
+
+### 5.1 需要迁移的代码
+
+goldfish.hpp 中 os 相关代码分布在行 264–268、370–726，具体包括：
+
+**辅助函数**：
+- `string_vector_to_s7_vector`（行 117–125）— 在 liii_os.cpp 中重复定义
+- `tb_directory_walk_func`（行 595–603）— 目录遍历回调
+
+**os 函数列表**：
+- `f_os_arch` / `glue_os_arch`（行 370–379）
+- `f_os_type` / `glue_os_type`（行 381–400）
+- `f_os_call` / `glue_os_call`（行 402–437）
+- `f_subprocess_run_values` 前向声明 + `glue_subprocess_run_values`（行 439–447）
+- `f_system` / `glue_system`（行 449–461）
+- `f_access` / `glue_access`（行 463–484）
+- `f_set_environment_variable` / `glue_setenv`（行 486–499）
+- `f_unset_environment_variable`（行 264–268）+ `glue_unsetenv`（行 501–506）
+- `f_os_temp_dir` / `glue_os_temp_dir`（行 508–520）
+- `f_mkdir` / `glue_mkdir`（行 522–533）
+- `f_rmdir` / `glue_rmdir`（行 535–546）
+- `f_remove_file` / `glue_remove_file`（行 548–560）
+- `f_rename` / `glue_rename`（行 562–580）
+- `f_chdir` / `glue_chdir`（行 582–593）
+- `f_listdir` / `glue_listdir`（行 605–637）
+- `f_getcwd` / `glue_getcwd`（行 639–651）
+- `f_getlogin` / `glue_getlogin`（行 653–669）
+- `f_getpid` / `glue_getpid`（行 671–685）
+- `glue_liii_os` 入口函数（行 706–726）
+
+### 5.2 迁移方式
+
+参照 `src/liii_njson.cpp` 和 `src/liii_http.cpp` 的模式：
+
+1. **新建 `src/liii_os.cpp`**，放在 `namespace goldfish` 中
+2. 按需引入头文件（`s7.h`、`<tbox/tbox.h>`、`<filesystem>`、`<vector>`、`<string>` 等），不 `#include "goldfish.hpp"`
+3. 在 `namespace goldfish` 内重复定义 `glue_define` 辅助函数
+4. 在 `namespace goldfish` 内重复定义 `string_vector_to_s7_vector` 辅助函数
+5. 提供 `void glue_liii_os(s7_scheme* sc)` 入口函数，移除 `inline` 修饰
+6. 在 `goldfish.hpp` 中删除 os 实现代码，仅保留前向声明：`void glue_liii_os(s7_scheme* sc);`
+7. 在 `xmake.lua` 中添加 `add_files("src/liii_os.cpp")`
+
+### 5.3 注意事项
+
+- `f_subprocess_run_values` 的实现在 `src/liii_subprocess.cpp` 中，liii_os.cpp 只需保留前向声明
+- `f_unset_environment_variable` 在 goldfish.hpp 中的位置（行 264–268）与 `glue_unsetenv`（行 501–506）分离，迁移时需注意完整提取
+- `f_getlogin` 使用了 `<pwd.h>` 和 `<unistd.h>`，需在 liii_os.cpp 中包含对应的条件编译头文件
+- `f_os_call` 使用了 `<wordexp.h>`，需在 liii_os.cpp 中包含对应头文件
+- `f_rename` 使用了 `<filesystem>`，需引入 `std::filesystem`
+- `goldfish.hpp` 中 `glue_liii_os` 的调用位置（行 1474）不需要修改
